Code cleanup
Leveraging a Map to hold the logic to expand a particular node type.

Also using the spread operator to clearly and succinctly define the
order and number of child nodes for each type.

Performance impact appears negligible on Node 7.2.1.  It should be
faster when the latest spread operator optimizations from v8 make it
into a Node release.